+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2023-03-14
 ### Added
 - Select filter type with static and dynamic values
 - BREAKING: Support for Vega charts with the new `vega` library
@@ -99,7 +101,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release of `datasette-dashboards`
 
-[Unreleased]: https://github.com/rclement/datasette-dashboards/compare/0.2.2...HEAD
+[Unreleased]: https://github.com/rclement/datasette-dashboards/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/rclement/datasette-dashboards/compare/0.2.2...0.3.0
 [0.2.2]: https://github.com/rclement/datasette-dashboards/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/rclement/datasette-dashboards/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/rclement/datasette-dashboards/compare/0.1.6...0.2.0